last update (testing)

+++ BomberMan/Class/Items/Bomb/Fires.cs
@@ -15,7 +15,7 @@ namespace BomberMan
         }
         public void Up(Point bomb)
         {
-            
+
             fire = new PictureBox()
             {
                 Size = Storages.TileSize,
@@ -23,57 +23,62 @@ namespace BomberMan
                 Image = Images.Fire,
                 SizeMode = PictureBoxSizeMode.Zoom,
             };
+            Storages.Map.AddTiles(fire);
             Control Tempbox = new Control();
             bool delete = false;
+            bool CanCreateNewFire = true;
             Items Tempitem = new Items();
             bool deleteItem = false;
-            Storages.Boxs.ForEach((box) =>
-            {
-                if (box.Location == fire.Location)
-                {
-                    Tempbox = box;
-                    delete = true;
-                    Storages.Map.DeleteTile(box);
-                    Storages.Tiles.Remove(box);
-                    Storages.LocationItemRandom.Add(box.Location);
-                    RandomItems.Randomitem();
-                }
-            });
             Storages.Items.ForEach((item) =>
             {
                 if (item.Location == fire.Location)
                 {
                     Tempitem = item;
+                    CanCreateNewFire = false;
                     deleteItem = true;
                     Storages.Map.DeleteTile(item.Image);
                     RandomItems.Randomitem();
                 }
             });
             
+            if (!deleteItem) 
+            {
+                Storages.Boxs.ForEach((box) =>
+                {
+                    if (box.Location == fire.Location)
+                    {
+                        Tempbox = box;
+                        CanCreateNewFire = false; 
+                        delete = true;
+                        Storages.Map.DeleteTile(box);
+                        Storages.Tiles.Remove(box);
+                        Storages.LocationItemRandom.Add(box.Location);
+                        RandomItems.Randomitem();
+                    }
+                });
+            }
+            
+            Storages.Fires.Add(fire);
             if (deleteItem)
             {
                 Storages.Items.Remove(Tempitem);
                 Storages.Map.DeleteTile(Tempitem.Image);
             }
-            Storages.Fires.Add(fire);
-            Storages.Map.AddTiles(fire);
-            if (delete)
+            else if (delete)
             {
                 Storages.Boxs.Remove(Tempbox);
             }
-            if (Storages.Player.Power > 1 && !delete)
+            else if (Storages.Player.Power > 1 && CanCreateNewFire)
             {
                 Storages.Player.Power -= 1;
                 System.Console.WriteLine("test");
                 Up(fire.Location);
                 Storages.Player.Power += 1;
             }
-            
-
         }
         public void Left(Point bomb)
         {
-            
+
 
             fire = new PictureBox()
             {
@@ -120,7 +125,7 @@ namespace BomberMan
             {
                 Storages.Boxs.Remove(Tempbox);
             }
-            if (Storages.Player.Power  > 1 && !delete)
+            if (Storages.Player.Power > 1 && !delete)
             {
                 Storages.Player.Power -= 1;
                 Left(fire.Location);
@@ -130,7 +135,7 @@ namespace BomberMan
         }
         public void Down(Point bomb)
         {
-            
+
             fire = new PictureBox()
             {
                 Size = Storages.TileSize,
@@ -165,7 +170,7 @@ namespace BomberMan
                     RandomItems.Randomitem();
                 }
             });
-            
+
             if (deleteItem)
             {
                 Storages.Items.Remove(Tempitem);
@@ -177,7 +182,7 @@ namespace BomberMan
             {
                 Storages.Boxs.Remove(Tempbox);
             }
-            if (Storages.Player.Power > 1&&!delete)
+            if (Storages.Player.Power > 1 && !delete)
             {
                 Storages.Player.Power -= 1;
                 Down(fire.Location);
@@ -187,7 +192,7 @@ namespace BomberMan
         }
         public void Right(Point bomb)
         {
-            
+
             fire = new PictureBox()
             {
                 Size = Storages.TileSize,
@@ -223,7 +228,7 @@ namespace BomberMan
                         return;
                     }
                 });
-            
+
             if (deleteItem)
             {
                 Storages.Items.Remove(Tempitem);
@@ -235,7 +240,7 @@ namespace BomberMan
             {
                 Storages.Boxs.Remove(Tempbox);
             }
-             if (Storages.Player.Power > 1 && !delete)
+            if (Storages.Player.Power > 1 && !delete)
             {
                 Storages.Player.Power -= 1;
                 Right(fire.Location);
